requirements.txt
ccxt>=2.0.0
python-dotenv
pandas
